2nd commit

+++ Backend/src/main/java/com/DeliveryOnTimeBackend/Backend/repository/UserRepository.java
@@ -7,7 +7,7 @@ import com.DeliveryOnTimeBackend.Backend.model.user;
 
 
 @Repository
-public interface UserRepository extends JpaRepository <user,Integer>{
+public interface UserRepository extends JpaRepository <user,Long>{
